Proper caps on acronyms & names
This is a pet peeve of mine... trying to keep some consistency in the way acronyms and names are written. Here is the corrected form:

macOS
HTTP
API
REST
Pydio
MySQL

Note that this only applies to strings emitted to some output device, *not* code!

Also, vendors were *not* touched.